Add a feature for serde deny unknown fields
Currently there is no way to turn off the deny unknown fields attribute.
This can cause issues when using the crate in cases where there are new
fields added that are not yet in the types structs, e.g. in unmerged
Core PRs.

Add a feature to types that can be used to enable this attribute, which
is off by default.

Forward the feature to client and then to node.

Use the feature in integration_test, through the node dependency.